Ok so I reformatted the code. Yeah, I like my {} on separate lines. So that's the end of that discussion.
I added in the option to save usernames and passwords. These will go to the systems credential manager not into the files. Thanks @hermano for the idea.

I also added the clear cache button so if you need to clear out the saved .jar's then you can for each one just hit the button and then redownload them.